+++ src/main/java/racing/Main.java
@@ -2,6 +2,10 @@ package racing;
 
 import java.util.Scanner;
 
+import racing.domain.Cars;
+import racing.view.InputView;
+import racing.view.OutputView;
+
 public class Main {
 	public static void main(String[] args) {
 		InputView inputView = new InputView();
