+++ src/test/java/racing/InputViewTest.java
@@ -13,6 +13,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import racing.view.InputView;
+
 public class InputViewTest {
 	@Test
 	public void splitAsCommaTest() {
