+++ src/test/java/racing/OutputViewTest.java
@@ -12,6 +12,9 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import racing.domain.Car;
+import racing.view.OutputView;
+
 public class OutputViewTest {
 	@ParameterizedTest
 	@MethodSource("generateWinner")
